+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 ## [Unreleased]
 
+### Added
+ - **Strings:**
+   - Add the uniqueKey function
+
 ## [1.0.6] - 2025-09-27
 
 ### Added